chore: create a notepad editor project
Launched a development process of brand new,
Notepad project. This one will contain an editor
that will allow user to work on the text note and
applying some styles to the text or some separate
parts. That means: bold, italic, underline,
highlight and more. For now, the general,
simplified UI is created, waiting on some new
exciting changes to be applied in the near future.