feat: apply functionality for Notepad buttons
This update consists of a few changes, that
enables the full power of Notepad buttons. That
is, user can bold, italize, highlight text (and
much more!). This will also include the ability
to create clickable links that may redirect user
to a specific web resource. That part is still
in the process of making though.