feat: create proper linking mechanism
The "link" button was the main focus as of this
update. The linking functionality had some weird
tweaks that has to be fixed. The work involved
adding proper conditions to disable the button
from being used when f.e. no text was selected.
This, combined with some help from event listeners
made it possible, and the whole mechanism is
working fine.