feat: enable setting background and text color
This new update contains what another handy
feature, which is a coloring buttons. There are
two to be exact. One of them allows to style the
text color; the second one sets the background
color of the given text node. Both button enable
customization to its fullest extent, giving the
user freedom of choice across the entire color
palette.