feat: add font sizing functionality
This small update contains a custom input, that
is able to set the font size value for the text
inside the editor. There are also extra buttons
(-) and (+) that are to decrease and increse
the font size value by 1, respectively. Font size
value is expressed in pixels, and cannot be lower
than 6 pixels, or bigger than 96 pixels.