feat: add undo and redo buttons
- This update introduces another two buttons.  Those are the undo & redo
buttons (second row, third and fourth column respectively). They enable
to travel back and forth through the edit history of the document. The
history state is tracked for 100 most recent updates inside the editor.

- For the font size input, there is also a list of predefined font sizes
that can be quickly chosen. The list becomes active when font size input
is focused , and it disappears when either a font size is chosen from
this list or when there is a click outside of the list.