feat: add support for font family (font change)
This update allows to add a variety of fonts to
the working area. User can smoothly change the
fonts to be used in the document, by picking one
from the list of predefined, supported font
families. Currently there is a small subset of
available fonts, but it may change in the future.