feat: add text alignment and list creation buttons
This update contains 6 new buttons, that extend
editor functionality even further. Four of them
are the alignment buttons, that allow to position
the specific line of text towards left, right,
center or to justify the text. Another two
buttons are used to create lists - former is for
bullet list and the latter is for ordered list
(each list items preceded with a number).

Also added a placeholder text, that appears when
the editor has no content yet. The text is:
"Write content here...", but the placeholder
message may be changed in future updates.